fix: fold routeless fixtures into cross_path_fixture for pkg file-count
Addresses PR #648 blocking feedback: pkg/factory/validation exceeded the
15-file limit after routeless_output_routes_fixture.go was added.
